requirements.txt
fastapi
uvicorn
python-jose
python-multipart
passlib[argon2]
argon2-cffi
python-dotenv
